pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simcraft"
version = "1.0.0"
description = "A discrete event simulation framework for Python"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Bulent Soykan" , email = "soykanb@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = [
    "simulation",
    "discrete-event",
    "DES",
    "modeling",
    "optimization",
    "reinforcement-learning",
]
dependencies = [
    "sortedcontainers>=2.4.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
visualization = [
    "matplotlib>=3.5.0",
    "pandas>=1.4.0",
]
rl = [
    "torch>=2.0.0",
]
all = [
    "matplotlib>=3.5.0",
    "pandas>=1.4.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/bulentsoykan/simcraft"
Documentation = "https://simcraft.readthedocs.io/"
Repository = "https://github.com/bulentsoykan/simcraft"

[tool.setuptools.packages.find]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
